Integer por Long

+++ src/main/java/com/sistema/apicr7imports/controller/CountryController.java
@@ -49,7 +49,7 @@ public class CountryController {
 		    @ApiResponse(code = 404, message = "Pais não encontrado."),
 		})
 	@GetMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<Country> findById(@ApiParam(value = "ID de Cadastro no Banco.", required = true,example = "1") @PathVariable long id) {
+	public ResponseEntity<Country> findById(@ApiParam(value = "ID de Cadastro no Banco.", required = true,example = "1") @PathVariable Integer id) {
 		return ResponseEntity.ok().body(service.findbyId(id));
 	}
